requirements.txt
langchain~=0.0.144
requests~=2.28.2
python-dotenv~=1.0.0